Fixes for AJAX-loaded fields
Hooks "renderReadyHook" instead of "render" so that JS config is
populated for AJAX-loaded fields. As a consequence, minimum required PW
version is now 3.0.61.
Moves JS initialisation to a function. Function is called when image
field is 'reloaded' (fix for AJAX-loaded fields) and when images are
uploaded (fix for new uploads) as well as DOM ready.
JS checks for modified inputfield name when rendered inside a repeater
item.
Also fix for newline character in tags config field.